add unit testing via jest
