+++ .gitignore
@@ -9,3 +9,4 @@ index
 index.exe
 index-*
 dist/
+coverage/
