+++ babel.config.cjs
@@ -1,0 +1,1 @@
+module.exports = { presets: ['@babel/preset-env'] };
