+++ defaults.js
@@ -1,0 +1,9 @@
+const defaults = {
+  sourceLang: 'en',
+  targetLang: 'es',
+  supportedLangs: ['en', 'es'],
+  services: ['ibm', 'deepl'],
+  apiService: 'ibm'
+};
+
+export default defaults;
