+++ rollup.config.js
@@ -3,7 +3,7 @@ import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
 
 export default {
-  input: 'index.js',
+  input: 'bin.js',
   output: {
     dir: 'output',
     format: 'cjs'
